Fix compilation errors and warnings
- Resolve type ambiguity in Wire.requestFrom() by explicit casting.
- Fix String to const char* conversion and Flash string comparisons in telegram.cpp.
- Fix function ordering and remove duplicate definitions in server.cpp.
- Ensure formatTimeBuffer is defined before use.

+++ src/main.cpp
@@ -261,7 +261,7 @@ byte writePCF8574(byte data) {
 
 // Функция для чтения байта с PCF8574
 byte readPCF8574() {
-  Wire.requestFrom(PCF8574_ADDRESS, (uint8_t)1); // Запросить 1 байт данных
+  Wire.requestFrom((uint8_t)PCF8574_ADDRESS, (uint8_t)1); // Запросить 1 байт данных
   if (Wire.available()) {
     return Wire.read();
   } else {
